Create solution_bt.py
Backtracking solution with O(n^2*m) time complexity and O(n*m) space complexity.

Memoize the dfs iteration of going through all brute-force cases, and memoize each (i, m) pair, where i is the break at which to create the subarray partition, and m is the number of splits. At each iteration, besides covering the termination cases of m == 1 and if the value exists in the dp memoization array, update the current sum and max sum, and the result. If the current sum is greater than the result, break out in order to optimize.

Sometimes this passes on Leetcode, sometimes not