Create solution_bs.py
O(n log s) binary search solution using the max element as the left bound and the sum of all the elements as the right bound. O(1) space.

Set the result to the right bound (sum of elements) as the maximum possible value.

Define the function to check whether the current largest value can be used to split the array of numbers in the appropriate subarrays. This is done by iterating through each element in the array, adding to a running sum, and splitting if the running sum becomes larger than the largest possible value. Then, the current sum is set to this value and the number of subarrays is incremented. Then, whether the subarrays are able to be formed is returned.

In the binary search, the midpoint is found, then the ability to be split is checked, and the search space is adjusted by the pointer for the right or left pointer.

Eventually, return the result.