Create solution.py
O(n + m) time and O(1) space solution, where n is the length of list A and m is the length of list B.

While the pointers for list A and B are not the same, iterate and keep going. If the pointer is null, then go to the head of the other list. When they first intersect, it will exit out of the while loop (or be null).

Return the pointer to list A (doesn't matter if it is list B either, since the loop will exit when they are the same/null).