Create solution.py
O(n) time solution using fast and slow pointers and O(1) space.

First, use fast and slow pointers to reverse first half of the list. Then, the prev and slow pointers will be the left half reversed and the right half in the original order.

Next, iterate prev or slow pointer and update the sum for the result, and iterate the pointers. Return the result.