Create solution.py
O(n^2) time complexity solution with two pointers and O(1) space.

Similar to 3Sum, first sort the array (O(n log n) operation) then iterate through in double loop (O(n^2) operation). Do this by enumerating (skipping any values that are the same - minor optimization).

Set the left and right pointers to the rest of the values, and while the pointers don't cross, calculate the three-sum. If it is equivalent to the target, return immediately. Otherwise, calculate the absolute difference and update the result.

Finally, if the three_sum is less than the target increment the left pointer, otherwise decrement the right pointer.

Return the result.