Create solution.py
Pointer manipulation solution in O(n) time and O(1) space.

Return none in the case nothing is in the list. Otherwise, set the odd pointer to the head, and the even to the next pointer from head. Also, set the even head to even.

While even and the next pointer of even exists, set the next pointer of odd to the next pointer of even (essentially odd.next.next). Then, update the odd pointer to its next pointer (advancing the newly set odd.next to be the current odd pointer). Then, set the next pointer of even to the next pointer of odd (the value after, even.next.next). Then, update the even pointer to the next pointer.

Finally, set the odd pointer's next value to the even pointer's head. Return the head