Create solution.py
O(n) time and O(1) space sliding window using left and right pointers.

First, take care of edge case where k <= 1.

Then, set the left and right pointers to the beginning. Initialize result to 0 and the current product to 1.

Then, while the right pointer is in bounds, multiply the right side to the current product. While this running product is greater than or equal to k, divide the left pointer element from it and increment the left pointer.

Update the result by 1 and the interval width of the left pointer to right pointer, and increment the right pointer.

Return the result