Create solution_arr.py
Dynamic array-based solution allocating array that resizes with every new webpage visited. This makes moving through the array much more efficient when using back() and forward() (both O(1) operations). However, visit() is a constant time operation when appending, but mimicking stack by not popping() but doing a soft-delete by just adding value to the next index of the array. Thus, this is just updating a pointer in the array and takes O(1) overall.