Create solution.py
Implementation of LFU cache using a doubly-linked list and hashmaps for O(1) amortized get() and put() operations. The pointer manipulation and using a dictionary achieves O(1) time for these operations. O(n) space is necessary for the dictionaries.

First, a ListNode class is defined in the standard way. Then, a LinkedList class is defined.  The constructor contains the leftmost and right most dummy nodes, as well as the map that maps the values of the ListNodes to the ListNodes themselves. There are a few different methods, including length() to return the length of the map (which in turn returns the length of the LinkedList), pushRight() to add a ListNode to the right of the list, while pop() and popLeft() pop from the right and left of the LinkedList, respectively. Meanwhile, update() pops and then pushes to the right.  This LinkedList class uses the dictionary to mimic the LRU cache doubly-linked list implementation with a dictionary for O(1) access and update.

Now, the LFUCache class is implemented using the LinkedList implementation. First, the constructor contains the capacity, as well as an lfu_cache for the least frequently used count, as well as dictionaries for mapping various values to keys of the frequency counts.

get() and put() are both O(1) owing to the dictionaries taking care of the operations in an amortized way, and the counter() is a way to increment count while adjusting the various dictionaries and counts.