Create solution_mergesort_bottom_up.py
O(n log n) time and O(1) space solution using bottom-up mergesort. The idea is to grab sublists of sizes of powers of 2 until they merge in a bottom-up manner.

First, initialize the step size to 1. Then, have a while loop that is true and evaluates until an return condition is met. This loop will record the progress of each pass into a single semi-sorted list by updating the next nodes of the first loop and dummy for the initial loop. In order to do this, first, initialize previous node to the dummy, then get the remaining as the next node of prev.

Initialize the number of loops to 0, and while there are remaining nodes, increment the number of loops. Create 2 sublists whose elements are step_size length away from the current node.  Then, iterate for each of the 2 sublists to get the remaining node with the substeps size being equalized to the step size.

While the substep size is not 0 and there are remaining nodes, decrease the substep size and set the tail of the corresponding sublist to the remaining node. Advance the next pointer for the remaining node.

If the sublist was filled during the iteration, make sure it terminates by checking that if the tail of a sublist exists, then put the next pointer to null (otherwise, this will be an infinite loop).

Now, while these two sublists are not empty, perform the merge operation by comparing each sublists' values and adjusting the pointers accordingly. Then, if either sublist still remains, append that to the end.

Double the step size at each iteration, and finally, check if the number of loops is down to 1 or less, which means the sorting is done. Return the head (dummy.next).