Create solution_mergesort_bottom_up_buffer.py
O(n log n) time complexity with O(1) size, same as the bottom-up mergesort approach, but with a holding buffer that speeds up some of the computation. The algorithm is the same as the bottom-up mergesort with constant space, but instead, the sublists and sublists_tail is of a buffer size (8 in this example), and the iteration for the sublists are done up to the buffer size.

For the merge step, the merge is done by going from 8 sublists, to 4 to 2 to 1. Thus, a sublist dummy is created, along with a sublist previous node and at each subsequent iteration, the number of sublists is halved.