Create solution.py
O(n) time and O(1) space solution. First set left and right nodes, and left tail and right tail initially as the dummy left and right nodes that were just set.

While the head is not empty, if the condition is met, set the current left tail to the current node, and update the left tail. Otherwise, set the right tail similarly. Each iteration, advance the head pointer.

Finally, set the left tail's next node to the next node of the right node (since right is just a dummy, and the actual value starts at right's next node), and set the right tail's next value to null (otherwise, it will point to another value potentially).

Return the next node of left.