Create solution.py
O(log n) time solution with O(1) space complexity, where the base of the log is 26 (the number of alphabet characters). Since the A characters corresponds to 1 (not 0), this is converting from base 10 to base 26 with a special condition of having A as 1 and not 0. Thus, the offset and integer division calculations are done with subtracting 1. Then, the result is appended by the character of the corresponding ord and offset values. Finally, the result is reversed (since the results were appended).