Create solution.py
O(min(m, n) + m +n) solution, where m is the length of string1 and n is the length of string2.

First, create a function to check if a substring is a divisor of both str1 and str2. If the modulus operation is not 0, then immediately return false. Otherwise, get the factors (length of str / the input parameter l) and return whether the substring multiplied by the factor is the same.

Now, iterate in a greedy manner starting from the minimum of the length of two strings all the way down to 0, and if it is a divisor, return that substring. Otherwise, at the end of the loop, return an empty string.