Create solution.py
Recursive O(h) solution, where h is the height of the BST (may not be balanced). Recursion is also O(h) space.

If the root is null, return it.

If the key is greater than the root value, assign the root's right value to the return value of deleting on the right side. If the key is less, then assign the root's left to the deleting operation on the left side.

Otherwise, if the left subtree is empty, return the right subtree and vice versa. In the case where both subtrees exists, set a current pointer to the right subtree of the root and keep going left to find the minimum value of the right subtree. Assign this value to the root value, then set the right subtree to the recursive step of deleting the current value.

Finally, return the root.