Create solution_bfs.py
O(n) time BFS level-order solution with O(n) space for the queue.

Traverse each level and get the length of the queue while queue is not empty (in order to see when to terminate each level).  Then, if the rightmost element at that level is not null, append to the result.