Create solution_dfs.py
O(n) time DFS solution with O(n) from recursion.

First, create an empty result list and set level to 0.

Define a helper function that will be recursively called, with base case for when the root is null. If the length of the result array (on the same level), then append the value for the current node. Perform DFS on the right child first, and pass in the incremented level, and then perform DFS on the left child.

Finally, perform DFS at the root and initialized level, and return the result.