Create solution.py
O(n) time and O(n) space for DFS traversal and adjacency matrix as well as recursion.

First, construct the adjacency matrix using the edges list of lists.

Then, define the DFS function with time initialized to 0. For each child in the adjacency matrix for the current node, if the child is equal to the parent, skip to the next iteration. Then, calculate the child time by recursion using DFS with the current node as the parent. If the child time is not 0 or the child has an apple, then add the child time + 2 (since the roundtrip adds 2 steps) to the time

Return the value of the DFS starting at 0 with parent -1 (just a dummy value).