Create solution.py
O(n) time and O(n) space BFS solution using queue. First, create a result array and deque with root if root exists, otherwise an empty deque.

Then, while the queue is not empty, create a level list and append the values by popping from the queue. Reverse the level if it is an odd index level, starting from level 0.

Return the result