Create solution.py
O(n^2) solution with O(n) space for the hashmap.

First, initialize the result to 1 (since there is at least 1 point).

Now, for each of the points, create a dictionary of counts. Then, for each of the other points available, find the slope. If the x values are the same, set slope to infinity. Otherwise, calculate slope as (y2 - y1)/(x2 - x1). 

Add this to the count with the key using the slope and increment. it in the dictionary.

Finally, adjust result to the max of the current result or the count + 1 for the current slope.

Return the result