Create solution.py
O(n^2) time solution with O(1) space for in-place matrix rotation.

First, set left and right pointers of matrix and then while the pointers don't cross, go through range of pointer (r - l) and set the top and bottom pointers to the same value as l and r (since this is a square matrix).

Then, rotate the edges of the square counterclockwise, saving the top left value. Appropriately adjust indices by adding the i index to the left of top-left, subtracting i from bottom of bottom-left, subtracting i from the right of bottom-right, and adding i to top of top-right. Decrement the r pointer and increment the i pointer each iteration.