Create solution.py
O(mn) solution where m and n are the number of rows and columns of the matrix, with O(1) extra space (besides output).

set left and right pointers and top and bottom pointers, with the bottom and right pointers being the length of the row and column.

While the pointers don't cross, go through in the order of top row left to right, right column top to bottom (and check here for pointers not crossing in edge cases like single row or single column matrices), then bottom row from right to left, then leftmost column from bottom to top. Adjust the pointers appropriately at each step.

Then return the result.