Create solution.py
O(mn) time solution with O(1) space in-place operation. Save the first column and first row for keeping track of whether to zero out those corresponding rows/columns, with another boolean variable to keep track of the first column's zero status.

Then, go through the list and record on the first column and first row whether to 0 or not.

In the second iteration, fill all non-first rows and columns with 0's according to the presence of 0 in the first row and columns.

Then, for the first column and row, zero out those rows/columns.