Create solution_math.py
Math solution using sum of squares and set for visited. O(n) space and time where n is the number of possible iterations of squaring and summing each digit.

Define the visited set and sum of squares function. If n is 1, return True immediately. Otherwise, add the number and update the number using the sum of squares of the digits.

If the loop exits (when a visited node is returned), return False.