Create solution_cycle_detection.py
Use fast and slow pointer cycle detection algorithm for linked lists. Apply the sumOfSquares twice for the fast pointer, and once for the slow pointer. This needs to be done ahead of the while loop (whose termination condition would be met if starting both at n, which makes the slow and fast pointers equal.

Finally, check if the slow pointer is 1, then return true, otherwise false. This algorithm is faster than the set or hashtable methods.