Create solution.py
O(n) time and O(1) space solution. Reversed digits array, set carry to 1 and the index to 0. While the carry is not 0, keep adding. If there is a 9, then set that digit to 0 and keep the carry at 1. Otherwise, increment and set carry to 0. Append the digit if the index is out of bounds and set carry to 0. Increment the i value.

Return the reversed digits array to undo the initial reversal.