Create solution.py
Maximum O(log_2(n)) solution for dividing by 2 each stage. The maximum number of times is 2^32 integer, which is 32 times.

Then, if n is less than or equal to 0 return false. Otherwise, for prime factors in 2, 3, 5, keep dividing by p and updating the number.

Return whether the number equals 1 at the end.