Create solution.py
O(n) time with n being the number of roman numeral characters in the string, and O(1) space (constant space for the hashmap).

Since the input is not going to be in bad format, check (if there exists a next character) whether that value is greater than the current value. Then subtract it from the result. Otherwise, add to the result.