Create solution.py
O(log(n)) time and O(1) space solution, where n is the number of times the values divide into the number due to repeated division.

For each of the Roman numeral and corresponding values, starting from the largest value, keep dividing and get the count. Append the number of times the corresponding Roman numeral divides into the value, and then mod the number by this value to get the remainder. Append to the result.