Create solution.py
O(mn) time solution with O(m+n) space complexity, where m is the number of characters (digits) in num1 and n is the number of characters in num2.

First, if either value is 0, return 0 immediately.

Allocate an array of the size equal to the length of m + n (hence the O(m+n) space). Also, reverse both num1 and num2.

Iterate through each number digit by digit in reverse order, and calculate the multiplication. Add to the digit, then put the carry by dividing by 10 to the next index. Update the current position value by modding.

Then, reverse the digit and start the beginning digit and count until it is not a 0. Then, map the string type to each of the integers in the array, then join the array together and return.