Create solution.py
O(n) time for count() method and O(1) for all other methods. O(n) space to maintain list of points and the hashmap for points.

For count() method, initialize the resulting count to 0 and get the x and y coordinates of the point input parameter. Then, for x and y in the list, if the x and y distance is not equal or the x or y coordinates are on the same point, continue to the next iteration (since this means there is no square possible).

Then, add the number of counts of the cross points available (x, p_y) and (p_x, y) which would be the other points of the square. If these are not available, one of the values in the defaultdict would be 0.

Return the result.