Create solution.py
O(s) solution where s is the sum of the n values (the missing sum).

First, calculate the missing sum (the sum of the n values). Then, if it is greater than 6 times the number of n values or less than n, return an empty array (since it means n * 6 dice rolls is exceeded, or it is lower than n * 1 dice rolls).

Find the integer division of the target value by dividing this missing sum by n. If the target is equal to the regular (floating-point) division, then return an n-array of this target value. Otherwise, there is still some remaining value to be added.  Find the remaining portion of the sum then set initial index to 0. Then, while there is a remaining portion and the index is not out-of-bounds, add the minimum of 6 minus the value at the current index in the result array, or the remaining value. This is a greedy approach. Then, add this to the result array at that index, while simultaneously subtracting this from the remaining value and advancing the index.

Return the result.