Create solution.py
O(n^2 log n) time solution due to n^2 iteration at each level of the tree, which has height log(n) since it is balanced. Memory is due to recursive stack the same.

Use DFS to recursively go through and first set a boolean variable if all values are the same. If not, set this value to false and break out. By convention, set the top left of the submatrix in question as the index of the corresponding submatrix for each node.

If they are all the same, set to a leaf node and return the node. Otherwise, divide the search length by 2 and get appropriate top-left positions of the top-left, top-right, bottom-left, and bottom-right submatrices and perform DFS on each of them recursively.

Return the node with the 4 children.

Finally, outside the inner DFS function, perform DFS for the length of the grid with the origin (0, 0) of the matrix and return this value.