Create solution.py
O(n^2) time and O(n^2) space solution using preorder DFS traversal.

The O(n^2) time complexity comes from going through all n nodes, with each serialization step using preorder DFS taking O(n) time. The O(n^2) space solution is due to the hashmap of the serialized hashmap of n nodes, with potentially the length of the list stored as the value of each of the n nodes being potentially a list of length n.

The serialization during preorder traversal is done by using commas (,) as delimiters, with the value of 'None' as a default string in the base case. The subtrees' serialized strings are stored as keys in the hashmap, with a check if there already exists the key, then the result is appended (since it is a duplicate).

The preorder starts at the root, then the result array that is constructed is returned.