Create solution_monostack.py
O(n) time and O(n) space solution using monotonic stack.

Go through the linked list and maintain a monotonic stack. When the value at the current node is greater, then keep popping from the stack.

Create a dummy node, and while the stack is not empty, keep appending to this linked list and return.