Create solution_reverse_ll.py
O(n) time and O(1) space solution using reversing the linked list.

First, reverse the linked list using a function to start from the end of the linked list.

The current maximum value is set to the value at the end of the linked list, then while current's next pointer node is not empty, if the current max is greater, skip that node. Otherwise, update the current max and increment the pointer.

Finally, reverse the list (since it was reversed) and return.