Create solution.py
O(n) time and O(n) space BFS level-order traversal solution.

Create a queue initialized with the root.

While the queue is not empty, pop and assign the node. If the node is not empty, append its left and right children to the queue. Otherwise, while the queue is not empty, pop. If the popped node is not null, then return False (since all nodes should be null at the bottom level to be a complete binary tree).

Outside of the loop, return true.