Create solution.py
O(n) time and O(n) space solution.

First, initialize a dictionary with the values in the inorder array (since they are unique) as keys with the corresponding indices as the values. This will allow for an O(n) time solution (as opposed to an O(n^2) solution when searching for the index in the inorder array as well as slicing the array.

Define a helper function, passing in the leftmost and rightmost indices of the inorder array to focus on. Then, if the pointers cross, then return None. Otherwise, get the root from the postorder array (since it is the rightmost value in the postorder array), then find the index based on the root value from the hashmap. Similarly, recursively call on the left and right indices portion with relation to the index from the inorder values.

Finally, call the helper function with the endpoints as the indices.