Create solution.py
O(n) time and space BFS solution.

create a tuple representing the node in the binary tree, the number (similar to an index representing the value of a heap, and the level. Also, set the previous level to 0 and previous number to 1.

While the queue is not empty, get the node, number, and level by popping from the queue. If the level is greater than the previous level, adjust the previous level and previous number accordingly.

Calculate the result by the maximum of itself and the width of the current level (which is calculated by the number minus the previous number plus 1, which gives the width at that level.

If the left or right node exists, append the left (or right) node, with double the amount of the number (or 2 x number + 1 for the right node), and also increment the level.

Finally, return the result.