Create solution.py
O(n) time and space solution using BFS for tree traversal.

Create an adjacency matrix by using a defaultdict of lists. For each value in the range n, append for each manager their employees.

Initialize the queue with the headID and time 0, and the result to 0.

While the queue is not empty, get the employee ID and the time by popping from the queue. Then, update the result to the maximum of the current result and the time it takes for the current employee. For the employee in the adjacency matrix for the values for the current employee, append to the queue the employee ID of these employees and the sum of the current time and informing time.

Return the result.