Create solution.py
O(n) time and O(1) space (not considering input cost array).

First, append a 0 to the cost array (the top of the stairs index, whose cost to itself would be 0).

Then, iterate from the 3rd-to-last element (last element is 0 by default, and the 2nd-to-last element already is at the right value) to the first element at index 0. Calculate the cost as the minimum of the cost of 1 or 2 jumps, then add it to the current index value.

Finally, return the minimum of the first 2 index elements of the cost array. This is possible since it is guaranteed that the cost array would contain at least 2 elements.