Create solution.py
Min heap solution with O(k) memory for the min heap which contains the k largest elements (the minimum of which is the k-th largest element) that is maintained.

The constructor is potentially O(n log n), since the heap-building operation takes O(n) time, and the popping takes O(log n) for a total of (n - k) operations, which is O((n-k) log n), for a worst case of O(n log n).

For the add method, the heap push operation takes O(log n), with the pop operation also taking O(log n). Finally, the returning of the minimum value is an O(1) operation.