Create solution.py
O(n log n) time solution with O(n) for heapify (to build the heap) then popping from heap log(n) for n times through the heap (for a total of n log n). The space complexity is O(n) for the heap. The heap is a max heap (which is mimicked in Python by multiplying all values by -1 since Python only has a min heap by default).

After creating a heap representation of the negated values of the stones, while the heap has more than 1 element remaining, get the heaviest 2 stones (the most negative). Then, if the second (lighter) stone is greater (lighter) than the first, push to the heap the value of their difference.

Finally, in order to handle the edge case when nothing remains in the heap, append a 0 to the heap. Finally, return the absolute value of the minimum value (top value) of the heap.