Create solution.py
O(n) time and O(h) space solution for recursive preorder DFS solution, where n is the number of nodes, and h is the height of the tree (which contributes to the space due to recursion). On average, h = log(n) but may be at worst case n if the tree is a linked-list-like structure.

First, create a helper DFS function. If the node is empty, return 0. Otherwise, set the result to 1 if the current node's value is greater than or equal to the max value, otherwise 0. Update the max value. Also, add both the DFS results from the left and right children, passing in the updated max value. Finally, return the result.

At the top level, return the DFS performed on the root node with the root value as the starting max value.