Create solution_heap.py
O(n) space solution for heap with O(log n) for push and pull operations.

Keep a defaultdict of count dictionaries, a heap, and the index in the constructor.

For the push operation, increment the index and the count of the value in question. Then, push into the heap the negation of the count and the negation of the index (in order to keep track of the most frequent counts and the index), along with the value in question. This is done to keep a max heap.

When popping, get the value and decrement from the count dictionary the value count. Return the value.