Create solution_heap.py
O(n + n log k) = O(n log k) time complexity solution with O(n) space for the heap and dictionary.

Create a count dictionary for each value. Subsequently, create a heap using a tuple of the count and key, and heapify this (O(n) operation).

Then, get the k-largest elements in the heap (O(n log k) operation). Then, return the n for the values (not the counts themselves).