Create solution_str.py
Solution using string function 'lstrip()' to strip from the left any whitespace characters. O(n) time complexity for running through the string.

First, use 'lstrip()' to strip any whitespace characters from the string. Define the MAX_INT to 2^31 - 1 and MIN_INT to - 2^31.

If the string is empty, return 0.

Set the sign to -1 if there is a '-' character, and 1 otherwise by default. Similarly, set the starting index to 1 if there was a sign at the beginning of the string or 0 otherwise.

Set the result to 0.

While the string still has characters, assign the value at the index to a temporary variable. If it is not a digit, break out of the string. Otherwise, multiply the result by 10 and add the integer value of the temporary digit variable, and increment the index. Finally, multiply the sign to the result, and set it to the maximum of itself and the MIN_INT, then the minimum of itself and the MAX_INT and return.