Create solution_dfa.py
O(n) solution using DFA (deterministic finite automaton). O(1) space.

Assign state of 0 for initial state as well as any whitespaces encountered. The state of 1 is assigned from the beginning and any whitespaces that follow if there is a '-' or '+' character encountered. Subsequently, state of 2 is for any digits that follow from state of 0 or 1. Any other state is a final state where the loop is broken out of.

Go through each state, and transition to the appropriate state. Assign the sign and digit values to the result correctly, then return the value based on the limits and the sign.