Create solution_no_str.py
O(n) time solution and O(n) space using no string functions. The ord() function is used to calculate the character ASCII value and subtract for each digit relative to ord('0'). The string is sliced after using the whitespace strip function that is manually implemented.

The appropriate preprocessing of the string indices for the sign of '-' and '+' is performed, then when looping through the string, if the sign and maximum and minimum 32-bit integer values are encountered, then the result is immediately returned. Otherwise, the result multiplied by the sign is returned if a non-digit character is first encountered.

Outside the loop, the result multiplied by sign is returned (if all characters are digits).