Create solution.py
O(n) time and space using recursive DFS.

Validate the node with the left and right boundaries as the limits. Define the validate recursive DFS solution to return true if node is null. Otherwise, if the value of the node is not within the left and right boundaries, return false. Finally, return validate on the left child using the existing left boundary and the current node's value (since left child has to be less than the current node's value), and the right child to be within the current node's value and the right boundary (since the right child has to be greater than the current node). Return the logical AND operation of these two recursive DFS branches.

Finally, return the validate function result on the root node with the initial boundaries as -inf and inf.