Create solution.py
O(n) time and O(n) space using iterative inorder DFS.

Set counter n to 0 and initialize a stack. Set the current pointer to the root.

While the current pointer or stack is not empty, loop for the current pointer and append to the stack, and set current pointer to its left child.

Pop from the stack when that iteration finishes, and set current to the popped value. Increment n by 1 since this means the next value is found.

If n equals k, return the current node's value.

Otherwise, set current pointer to its right child. This loop will always finish since it is guaranteed there is at least k nodes in the tree.