Create solution.py
O(n) time and O(n) space solution due to recursion. 

First, for the base case, if the preorder or inorder lists are empty, return null.

Set the root as the 0th index of the preorder array. Find the mid index of the inorder array as the index of the preorder root (the value at the 0th index in the preorder). This is the splitting point for the inorder array.

Now, perform the buildTree recursion on the 1 to mid index on the preorder array, and up to (but not including) the mid index for the inorder array and assign it to the left of the root. The preorder is done from 1 to mid since these are the number of values that will be on the left side. Similarly, perform buildTree on the rest of the preorder array (from mid + 1 index to the end, and inorder on the mid + 1 index to the end and assign to the right subtree of the root.

Finally, return the root.