Create solution_dp.py
O(n^2) time and O(n) space dynamic programming solution. The key is to notice that we choose for each value from 1 to n one of these values as the root, then all of the other nodes are either to the left (or the right) of it. This is the basic idea of this, and all subtrees are calculated prior (memoized).

First, initialize the DP array for num trees, including the 0-th node, with values of 1. The base cases are n = 0, which corresponds to a value of 1 tree possible, as well as the n = 1 case, which can only form 1 tree.

Now, iterating from 2 to n, set the total to 0. Then, for the roots at the nodes from 1 up to the node value itself, calculate left as the root - 1, and right as the current nodes - root. Then, the total is the product of the number of trees possible in the left and right subtrees. Add this to the running total. Finally, set the DP array at the nodes value to the total just calculated.

Finally, return the value of the DP array for the index n, which contains the value of the number of possible subtrees.