Create solution.py
O(4^n/(n sqrt(n)) time complexity (according to Leetcode editorial) - this is exponential time complexity. The enumeration of all trees will be exponential as the number of trees grows. The space using a DP array (not including the result) allows for duplicate work to be minimized.

First, initialize the DP array.

Create a function to generate the number of trees based on the left and right bounds. If the left and right bounds cross, then return a list with a null value.

If the tuple of the left and right bound already exists in the DP array, return that value.

Initialize the result list.

For each value in the range of the left and right boundaries, iterate through each pairwise combination of the left and right subtrees by calling the generate function recursively, which are in the range of [left_bd, val) and (val, right_bd], where the parentheses and brackets indicate open and closed intervals. Now, set the node to the current value, with the left and right subtrees and create a new node, and append this to the result list.

Finally, add this node generated from the left and right bounds as the key to the DP array, and return the result.

Finally, return the generate function output from the range 1 to n.