Create solution.py
Recursive preorder DFS solution - O(n) time and O(h = log n) space for recursive call stack.

First, define the DFS function passing in as arguments the node and the current running sum. If the node is null, return 0.

Update the sum by multiplying by 10 and adding the value at the current node. If both the left and right children are null, return the current running sum. Otherwise, return the sum of the left child and right child, passing in the current sum.

Finally, run DFS at the root with initial value of 0 passed in as the current sum.