Create solution.py
O(n) time and O(1) space solution.

The key is to notice that if the 1st house is robbed, then the next house is skipped and the remainder's sum is looked at, while skipping the 1st house means looking at the max possible sum of the 2nd house to the end.

Since the value checked is only the previous 2 values, the problem only requires 2 variables to store the maximum possible up to that point. Thus, initialize two sum variables to 0. 

Then, for each number in the array, calculate the max  possible up to that point by adding the value to the sum1, vs. sum2 (which means the current value is skipped). Next, set sum1 to sum2 (since the next iteration will replace this point), and set sum2 to the value just calculated previously.

Finally, return the sum2 which contains the last value calculated.