Create solution.py
O(n) time solution and O(1) space. Same as House Robber I, re-using the same algorithm and applying it to two of the subarrays with either end of the array missing.

Apply the function to the array excluding either the leftmost or rightmost array, and take the max of these values, along with the first element of the array (in case the array only has 1 element).