Create solution_dfs.py
O(n) time and O(h = log(n)) space complexity using recursive DFS.

Define a modified DFS algorithm. If the node is null, return a tuple of (0, 0). Find the DFS of both the left and right subtrees then calculate the value with the root (in which case the value of the root is added, along with the value of not including the left and right nodes, the index-1 values of each), and the value without the root (which is just the max of the left and right subtrees). Return the tuple.

Finally, run the DFS on the root node and return the max of the tuple.