Create solution.py
O(n) time recursive DFS solution with O(h = log(n)) space due to recursive stack.

Define DFS function. If the node is null, return null. If the node is either equal to p or q, return the node.

Find the left and right side nodes using the same DFS for the left and right subtrees.

If both the left and right return a node, return the node itself (since this is the LCA, given that the previous condition of node == p or node == q was not triggered). Otherwise, return either of the left or right node, whichever is not null.

Finally, run DFS on the root and return this value.